bb-imager-gui: Cleanup code
- Just some refactor to improve code quality
