+++ bb-imager-gui/src/main.rs
@@ -1,11 +1,7 @@
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
-use std::{
-    collections::HashSet,
-    time::{Duration, Instant},
-};
+use std::time::Duration;
 
-use bb_config::config;
 use constants::PACKAGE_QUALIFIER;
 use iced::{Subscription, Task, futures::SinkExt, widget};
 use message::BBImagerMessage;
@@ -13,10 +9,13 @@ use tokio_stream::StreamExt as _;
 use tracing::level_filters::LevelFilter;
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
 
+use crate::state::BBImagerCommon;
+
 mod constants;
 mod helpers;
 mod message;
 mod persistance;
+mod state;
 mod ui;
 mod updater;
 
@@ -63,7 +62,7 @@ fn main() -> iced::Result {
     };
 
     iced::application(BBImager::new, message::update, ui::view)
-        .title(app_title)
+        .title(helpers::app_title)
         .subscription(BBImager::subscription)
         .theme(BBImager::theme)
         .window(settings)
@@ -87,747 +86,26 @@ fn main() -> iced::Result {
         .run()
 }
 
-fn app_title(_: &BBImager) -> String {
-    if option_env!("PRE_RELEASE").is_some() {
-        format!("{} (pre-release)", constants::APP_NAME)
-    } else {
-        format!("{} v{}", constants::APP_NAME, env!("CARGO_PKG_VERSION"))
-    }
-}
-
-#[derive(Debug)]
-struct BBImagerCommon {
-    app_config: persistance::GuiConfiguration,
-    boards: helpers::Boards,
-    downloader: bb_downloader::Downloader,
-    timezones: widget::combo_box::State<String>,
-    keymaps: widget::combo_box::State<String>,
-
-    // Constant image handles
-    board_svg_handle: widget::svg::Handle,
-    downloading_svg_handle: widget::svg::Handle,
-    arrow_forward_svg_handle: widget::svg::Handle,
-    format_svg_handle: widget::svg::Handle,
-    file_add_svg_handle: widget::svg::Handle,
-    arrow_back_svg_handle: widget::svg::Handle,
-    usb_svg_handle: widget::svg::Handle,
-    file_save_icon: widget::svg::Handle,
-    info_svg_handle: widget::svg::Handle,
-    window_icon_handle: widget::image::Handle,
-
-    img_handle_cache: helpers::ImageHandleCache,
-}
-
-impl BBImagerCommon {
-    fn updater_task(&self) -> Task<BBImagerMessage> {
-        if cfg!(feature = "updater") {
-            let downloader = self.downloader.clone();
-            Task::perform(
-                async move { updater::check_update(downloader).await },
-                |x| match x {
-                    Ok(Some(ver)) => BBImagerMessage::UpdateAvailable(ver),
-                    Ok(None) => {
-                        tracing::info!("Application is at the latest version");
-                        BBImagerMessage::Null
-                    }
-                    Err(e) => {
-                        tracing::error!("Failed to check for application update: {e:?}");
-                        BBImagerMessage::Null
-                    }
-                },
-            )
-        } else {
-            Task::none()
-        }
-    }
-
-    fn fetch_images(&self, iter: impl IntoIterator<Item = url::Url>) -> Task<BBImagerMessage> {
-        let tasks = iter.into_iter().map(|icon| {
-            let downloader = self.downloader.clone();
-            let icon_clone = icon.clone();
-            let icon_clone2 = icon.clone();
-            Task::perform(
-                async move { downloader.download_no_cache(icon_clone, None).await },
-                move |p| match p {
-                    Ok(p) => BBImagerMessage::ResolveImage(icon_clone2, p),
-                    Err(_) => {
-                        tracing::warn!("Failed to fetch image {}", icon);
-                        BBImagerMessage::Null
-                    }
-                },
-            )
-        });
-        Task::batch(tasks)
-    }
-
-    fn fetch_board_images(&self) -> Task<BBImagerMessage> {
-        // Do not try downloading same image multiple times
-        let icons: HashSet<url::Url> = self
-            .boards
-            .devices()
-            .filter_map(|(_, dev)| dev.icon.clone())
-            .collect();
-
-        self.fetch_images(icons)
-    }
-
-    fn fetch_os_images(&self, board: usize, target: &[usize]) -> Task<BBImagerMessage> {
-        let Some(os_images) = self.boards.images(board, target) else {
-            return Task::none();
-        };
-
-        // Do not try downloading same image multiple times
-        let icons: HashSet<url::Url> = os_images.map(|(_, x)| x.icon()).cloned().collect();
-
-        self.fetch_images(icons)
-    }
-
-    fn fetch_remote_subitems(&self, board: usize, target: &[usize]) -> Task<BBImagerMessage> {
-        let Some(os_images) = self.boards.images(board, target) else {
-            // Maybe resolving was missed
-            if let config::OsListItem::RemoteSubList(item) = self.boards.image(target) {
-                let url = item.subitems_url.clone();
-                tracing::debug!("Downloading subitems from {:?}", url);
-
-                let target_clone: Vec<usize> = target.to_vec();
-                let downloader = self.downloader.clone();
-
-                return Task::perform(
-                    async move { downloader.download_json_no_cache(url).await },
-                    move |x| match x {
-                        Ok(item) => BBImagerMessage::ResolveRemoteSubitemItem {
-                            item,
-                            target: target_clone.clone(),
-                        },
-                        Err(e) => {
-                            tracing::warn!("Failed to download subitems with error {e}");
-                            BBImagerMessage::Null
-                        }
-                    },
-                );
-            } else {
-                return Task::none();
-            }
-        };
-
-        let remote_image_jobs = os_images
-            .filter_map(|(idx, x)| {
-                if let config::OsListItem::RemoteSubList(item) = x {
-                    tracing::debug!("Fetch: {:?} at {}", item.subitems_url, idx);
-                    Some((idx, item.subitems_url.clone()))
-                } else {
-                    None
-                }
-            })
-            .map(|(idx, url)| {
-                let mut new_target: Vec<usize> = target.to_vec();
-                new_target.push(idx);
-
-                let downloader = self.downloader.clone();
-                let url_clone = url.clone();
-                Task::perform(
-                    async move {
-                        downloader
-                            .download_json_no_cache::<Vec<config::OsListItem>, url::Url>(url_clone)
-                            .await
-                    },
-                    move |x| match x {
-                        Ok(item) => BBImagerMessage::ResolveRemoteSubitemItem {
-                            item,
-                            target: new_target.clone(),
-                        },
-                        Err(e) => {
-                            tracing::warn!("Failed to download subitems {:?} with error {e}", url);
-                            BBImagerMessage::Null
-                        }
-                    },
-                )
-            });
-
-        Task::batch(remote_image_jobs)
-    }
-
-    pub(crate) fn image_handle_cache(&self) -> &helpers::ImageHandleCache {
-        &self.img_handle_cache
-    }
-
-    pub(crate) fn board_svg(&self) -> &widget::svg::Handle {
-        &self.board_svg_handle
-    }
-
-    pub(crate) fn downloading_svg(&self) -> &widget::svg::Handle {
-        &self.downloading_svg_handle
-    }
-
-    pub(crate) fn info_svg(&self) -> &widget::svg::Handle {
-        &self.info_svg_handle
-    }
-
-    pub(crate) fn window_icon(&self) -> &widget::image::Handle {
-        &self.window_icon_handle
-    }
-}
-
-#[derive(Debug)]
-struct ChooseBoardState {
-    common: BBImagerCommon,
-    selected_board: Option<usize>,
-}
-
-impl ChooseBoardState {
-    pub(crate) fn devices(&self) -> impl Iterator<Item = (usize, &config::Device)> {
-        self.common.boards.devices()
-    }
-
-    pub(crate) fn board_svg(&self) -> &widget::svg::Handle {
-        &self.common.board_svg_handle
-    }
-
-    pub(crate) fn downloading_svg(&self) -> &widget::svg::Handle {
-        &self.common.downloading_svg_handle
-    }
-
-    pub(crate) fn selected_board(&self) -> Option<&config::Device> {
-        Some(self.common.boards.device(self.selected_board?))
-    }
-
-    pub(crate) fn image_handle_cache(&self) -> &helpers::ImageHandleCache {
-        &self.common.img_handle_cache
-    }
-}
-
-#[derive(Debug)]
-struct ChooseOsState {
-    common: BBImagerCommon,
-    selected_board: usize,
-    pos: Vec<usize>,
-    selected_image: Option<(OsImageId, helpers::BoardImage)>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum OsImageId {
-    // Vec points to parent
-    Format(Vec<usize>),
-    // Vec points to parent
-    Local(Vec<usize>),
-    // Vec points to OsImage
-    Remote(Vec<usize>),
-}
-
-struct OsImageItem<'a> {
-    id: OsImageId,
-    icon: Option<&'a url::Url>,
-    label: &'a str,
-    is_sublist: bool,
-}
-
-impl<'a> OsImageItem<'a> {
-    fn format(parent: Vec<usize>, label: &'a str) -> Self {
-        Self {
-            id: OsImageId::Format(parent),
-            icon: None,
-            label,
-            is_sublist: false,
-        }
-    }
-
-    fn local(parent: Vec<usize>) -> Self {
-        Self {
-            id: OsImageId::Local(parent),
-            icon: None,
-            label: "Select Local Image",
-            is_sublist: false,
-        }
-    }
-
-    fn remote(id: Vec<usize>, url: &'a url::Url, label: &'a str, is_sublist: bool) -> Self {
-        Self {
-            id: OsImageId::Remote(id),
-            icon: Some(url),
-            label,
-            is_sublist,
-        }
-    }
-}
-
-impl ChooseOsState {
-    pub(crate) fn selected_image(&self) -> Option<(&OsImageId, &helpers::BoardImage)> {
-        match &self.selected_image {
-            Some((x, y)) => Some((x, y)),
-            None => None,
-        }
-    }
-
-    pub(crate) fn selected_board(&self) -> &config::Device {
-        self.common.boards.device(self.selected_board)
-    }
-
-    pub(crate) fn images(&self) -> Option<impl Iterator<Item = OsImageItem<'_>>> {
-        let iter = self
-            .common
-            .boards
-            .images(self.selected_board, self.pos.as_slice())?
-            .map(|(id, x)| {
-                let mut idx = self.pos.clone();
-                idx.push(id);
-
-                OsImageItem::remote(
-                    idx,
-                    x.icon(),
-                    x.name(),
-                    matches!(
-                        x,
-                        config::OsListItem::SubList(_) | config::OsListItem::RemoteSubList(_)
-                    ),
-                )
-            });
-
-        let extra = match self.flasher() {
-            config::Flasher::SdCard => vec![
-                OsImageItem::format(self.pos.clone(), "Format SD Card"),
-                OsImageItem::local(self.pos.clone()),
-            ],
-            _ => vec![OsImageItem::local(self.pos.clone())],
-        };
-
-        Some(iter.chain(extra))
-    }
-
-    pub(crate) fn image(&self, idx: &[usize]) -> &config::OsListItem {
-        self.common.boards.image(idx)
-    }
-
-    pub(crate) fn image_handle_cache(&self) -> &helpers::ImageHandleCache {
-        &self.common.img_handle_cache
-    }
-
-    pub(crate) fn flasher(&self) -> config::Flasher {
-        if self.pos.is_empty() {
-            self.selected_board().flasher
-        } else {
-            match self.image(&self.pos) {
-                config::OsListItem::Image(_) => panic!("Expected list"),
-                config::OsListItem::SubList(x) => x.flasher,
-                config::OsListItem::RemoteSubList(x) => x.flasher,
-            }
-        }
-    }
-
-    pub(crate) fn downloading_svg(&self) -> &widget::svg::Handle {
-        &self.common.downloading_svg_handle
-    }
-
-    pub(crate) fn arrow_forward_svg(&self) -> &widget::svg::Handle {
-        &self.common.arrow_forward_svg_handle
-    }
-
-    pub(crate) fn format_svg(&self) -> &widget::svg::Handle {
-        &self.common.format_svg_handle
-    }
-
-    pub(crate) fn file_add_svg(&self) -> &widget::svg::Handle {
-        &self.common.file_add_svg_handle
-    }
-
-    pub(crate) fn arrow_back_svg(&self) -> &widget::svg::Handle {
-        &self.common.arrow_back_svg_handle
-    }
-
-    pub(crate) fn downloader(&self) -> &bb_downloader::Downloader {
-        &self.common.downloader
-    }
-}
-
-#[derive(Debug)]
-enum DestinationItem<'a> {
-    SaveToFile(String),
-    Destination(&'a helpers::Destination),
-}
-
-impl<'a> std::fmt::Display for DestinationItem<'a> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            DestinationItem::SaveToFile(_) => write!(f, "Save To File"),
-            DestinationItem::Destination(d) => d.fmt(f),
-        }
-    }
-}
-
-impl<'a> DestinationItem<'a> {
-    fn msg(&'a self) -> BBImagerMessage {
-        match self {
-            DestinationItem::SaveToFile(x) => BBImagerMessage::SelectFileDest(x.clone()),
-            DestinationItem::Destination(d) => BBImagerMessage::SelectDest((*d).clone()),
-        }
-    }
-
-    fn is_selected(&'a self, dst: &'a helpers::Destination) -> bool {
-        match self {
-            DestinationItem::SaveToFile(_) => false,
-            DestinationItem::Destination(d) => dst.eq(d),
-        }
-    }
-}
-
-#[derive(Debug)]
-struct ChooseDestState {
-    common: BBImagerCommon,
-    selected_board: usize,
-    selected_image: (OsImageId, helpers::BoardImage),
-    selected_dest: Option<helpers::Destination>,
-    destinations: Vec<helpers::Destination>,
-    filter_destination: bool,
-}
-
-impl ChooseDestState {
-    fn selected_dest(&self) -> Option<&helpers::Destination> {
-        self.selected_dest.as_ref()
-    }
-
-    fn destinations<'a>(&'a self) -> impl Iterator<Item = DestinationItem<'a>> + 'a {
-        let iter = self.destinations.iter().map(DestinationItem::Destination);
-
-        let temp = match self.selected_image.1.file_name() {
-            Some(x) => vec![DestinationItem::SaveToFile(x)],
-            None => vec![],
-        };
-
-        iter.chain(temp)
-    }
-
-    fn usb_svg(&self) -> &widget::svg::Handle {
-        &self.common.usb_svg_handle
-    }
-
-    fn file_save_icon(&self) -> &widget::svg::Handle {
-        &self.common.file_save_icon
-    }
-
-    pub(crate) fn selected_board(&self) -> &config::Device {
-        self.common.boards.device(self.selected_board)
-    }
-
-    fn instruction(&self) -> Option<&str> {
-        match self.selected_image.1.info_text() {
-            Some(x) => Some(x),
-            None => self.selected_board().instructions.as_deref(),
-        }
-    }
-}
-
-#[derive(Debug)]
-struct CustomizeState {
-    common: BBImagerCommon,
-    selected_board: usize,
-    selected_image: (OsImageId, helpers::BoardImage),
-    selected_dest: helpers::Destination,
-    customization: helpers::FlashingCustomization,
-}
-
-impl CustomizeState {
-    fn timezones(&self) -> &widget::combo_box::State<String> {
-        &self.common.timezones
-    }
-
-    fn keymaps(&self) -> &widget::combo_box::State<String> {
-        &self.common.keymaps
-    }
-
-    fn customization(&self) -> &helpers::FlashingCustomization {
-        &self.customization
-    }
-
-    fn app_config(&self) -> &persistance::GuiConfiguration {
-        &self.common.app_config
-    }
-
-    fn save_app_config(&self) -> Task<BBImagerMessage> {
-        let config = self.app_config().clone();
-        Task::future(async move {
-            if let Err(e) = config.save().await {
-                tracing::error!("Failed to save config: {e}");
-            }
-            BBImagerMessage::Null
-        })
-    }
-
-    fn selected_board(&self) -> &str {
-        self.common.boards.device(self.selected_board).name.as_str()
-    }
-
-    fn selected_image(&self) -> String {
-        self.selected_image.1.to_string()
-    }
-
-    fn selected_destination(&self) -> String {
-        match self.selected_dest.size() {
-            Some(x) => format!("{} ({})", self.selected_dest, helpers::pretty_bytes(x)),
-            None => self.selected_dest.to_string(),
-        }
-    }
-
-    fn is_download(&self) -> bool {
-        self.selected_dest.is_download_action()
-    }
-
-    fn modifications(&self) -> Vec<&'static str> {
-        match self.customization() {
-            helpers::FlashingCustomization::LinuxSdSysconfig(x) => {
-                let mut ans = Vec::new();
-
-                if x.user.is_some() {
-                    ans.push("• User account configured");
-                }
-
-                if x.wifi.is_some() {
-                    ans.push("• Wifi configured");
-                }
-
-                if x.hostname.is_some() {
-                    ans.push("• Hostname configured");
-                }
-
-                if x.keymap.is_some() {
-                    ans.push("• Keymap configured");
-                }
-
-                if x.timezone.is_some() {
-                    ans.push("• Timezone configured");
-                }
-
-                if x.ssh.is_some() {
-                    ans.push("• SSH Key configured");
-                }
-
-                if x.usb_enable_dhcp == Some(true) {
-                    ans.push("• USB DHCP enabled");
-                }
-
-                ans
-            }
-            helpers::FlashingCustomization::Bcf(x) => {
-                if !x.verify {
-                    vec!["• Skip Verification"]
-                } else {
-                    Vec::new()
-                }
-            }
-            _ => Vec::new(),
-        }
-    }
-}
-
-#[derive(Debug)]
-struct FlashingState {
-    common: BBImagerCommon,
-    selected_board: usize,
-    cancel_flashing: iced::task::Handle,
-    progress: bb_flasher::DownloadFlashingStatus,
-    start_timestamp: Option<Instant>,
-    is_download: bool,
-}
-
-impl FlashingState {
-    pub(crate) fn selected_board(&self) -> &config::Device {
-        self.common.boards.device(self.selected_board)
-    }
-
-    fn time_remaining(&self) -> Option<Duration> {
-        const THRESHOLD: f32 = 0.02;
-
-        match self.progress {
-            bb_flasher::DownloadFlashingStatus::FlashingProgress(x)
-            | bb_flasher::DownloadFlashingStatus::DownloadingProgress(x) => {
-                if x < THRESHOLD {
-                    None
-                } else {
-                    let t = self.start_timestamp?.elapsed();
-                    let x = x.clamp(0.0, 1.0);
-                    let scale = (1.0 - x) / x;
-                    Some(t.mul_f32(scale))
-                }
-            }
-            bb_flasher::DownloadFlashingStatus::Customizing => Some(Duration::from_secs(1)),
-            _ => None,
-        }
-    }
-
-    fn progress_update(&mut self, u: bb_flasher::DownloadFlashingStatus) {
-        // Required for better time estimate.
-        match u {
-            bb_flasher::DownloadFlashingStatus::DownloadingProgress(_)
-            | bb_flasher::DownloadFlashingStatus::FlashingProgress(_) => {
-                if self.start_timestamp.is_none() {
-                    self.start_timestamp = Some(Instant::now())
-                }
-            }
-            _ => {}
-        }
-
-        self.progress = u;
-    }
-}
-
-#[derive(Debug)]
-struct FlashingFinishState {
-    common: BBImagerCommon,
-    selected_board: usize,
-    is_download: bool,
-}
-
-impl FlashingFinishState {
-    pub(crate) fn selected_board(&self) -> &config::Device {
-        self.common.boards.device(self.selected_board)
-    }
-}
-
-struct FlashingFailState {
-    common: BBImagerCommon,
-    err: String,
-    logs: widget::text_editor::Content,
-}
-
-// State for Pages that can be opened from any of the normal pages but are not part of normal flow.
-// Eg: Application info
-enum OverlayData {
-    ChooseBoard(ChooseBoardState),
-    ChooseOs(ChooseOsState),
-    ChooseDest(ChooseDestState),
-    Customize(CustomizeState),
-    Review(CustomizeState),
-    Flashing(FlashingState),
-    FlashingCancel(FlashingFinishState),
-    FlashingFail(FlashingFailState),
-    FlashingSuccess(FlashingFinishState),
-}
-
-impl OverlayData {
-    fn common_mut(&mut self) -> &mut BBImagerCommon {
-        match self {
-            Self::ChooseBoard(x) => &mut x.common,
-            Self::ChooseOs(x) => &mut x.common,
-            Self::ChooseDest(x) => &mut x.common,
-            Self::Customize(x) => &mut x.common,
-            Self::Review(x) => &mut x.common,
-            Self::Flashing(x) => &mut x.common,
-            Self::FlashingCancel(x) => &mut x.common,
-            Self::FlashingFail(x) => &mut x.common,
-            Self::FlashingSuccess(x) => &mut x.common,
-        }
-    }
-
-    fn common(&self) -> &BBImagerCommon {
-        match self {
-            Self::ChooseBoard(x) => &x.common,
-            Self::ChooseOs(x) => &x.common,
-            Self::ChooseDest(x) => &x.common,
-            Self::Customize(x) => &x.common,
-            Self::Review(x) => &x.common,
-            Self::Flashing(x) => &x.common,
-            Self::FlashingCancel(x) => &x.common,
-            Self::FlashingFail(x) => &x.common,
-            Self::FlashingSuccess(x) => &x.common,
-        }
-    }
-}
-
-impl TryFrom<BBImager> for OverlayData {
-    type Error = ();
-
-    fn try_from(value: BBImager) -> Result<Self, Self::Error> {
-        match value {
-            BBImager::ChooseBoard(x) => Ok(Self::ChooseBoard(x)),
-            BBImager::ChooseOs(x) => Ok(Self::ChooseOs(x)),
-            BBImager::ChooseDest(x) => Ok(Self::ChooseDest(x)),
-            BBImager::Customize(x) => Ok(Self::Customize(x)),
-            BBImager::Review(x) => Ok(Self::Review(x)),
-            BBImager::Flashing(x) => Ok(Self::Flashing(x)),
-            BBImager::FlashingCancel(x) => Ok(Self::FlashingCancel(x)),
-            BBImager::FlashingFail(x) => Ok(Self::FlashingFail(x)),
-            BBImager::FlashingSuccess(x) => Ok(Self::FlashingSuccess(x)),
-            BBImager::Dummy | BBImager::AppInfo(_) => Err(()),
-        }
-    }
-}
-
-impl From<OverlayData> for BBImager {
-    fn from(value: OverlayData) -> Self {
-        match value {
-            OverlayData::ChooseBoard(x) => Self::ChooseBoard(x),
-            OverlayData::ChooseOs(x) => Self::ChooseOs(x),
-            OverlayData::ChooseDest(x) => Self::ChooseDest(x),
-            OverlayData::Customize(x) => Self::Customize(x),
-            OverlayData::Review(x) => Self::Review(x),
-            OverlayData::Flashing(x) => Self::Flashing(x),
-            OverlayData::FlashingCancel(x) => Self::FlashingCancel(x),
-            OverlayData::FlashingFail(x) => Self::FlashingFail(x),
-            OverlayData::FlashingSuccess(x) => Self::FlashingSuccess(x),
-        }
-    }
-}
-
-struct OverlayState {
-    page: OverlayData,
-    log_path: String,
-    license: widget::text_editor::Content,
-    cache_dir: String,
-}
-
-impl OverlayState {
-    fn new(page: OverlayData) -> Self {
-        let log_path = helpers::log_file_path().to_string_lossy().to_string();
-        let license = widget::text_editor::Content::with_text(constants::APP_LINCESE);
-        let cache_dir = helpers::project_dirs()
-            .unwrap()
-            .cache_dir()
-            .to_string_lossy()
-            .to_string();
-
-        Self {
-            page,
-            log_path,
-            license,
-            cache_dir,
-        }
-    }
-
-    fn common(&self) -> &BBImagerCommon {
-        self.page.common()
-    }
-
-    fn common_mut(&mut self) -> &mut BBImagerCommon {
-        self.page.common_mut()
-    }
-
-    fn log_path(&self) -> &str {
-        &self.log_path
-    }
-
-    fn cache_dir_path(&self) -> &str {
-        &self.cache_dir
-    }
-}
-
 #[derive(Default)]
 enum BBImager {
     // Dummy state to allow clone-free move among variants. Should never be exposed in view.
     #[default]
     Dummy,
-    ChooseBoard(ChooseBoardState),
-    ChooseOs(ChooseOsState),
-    ChooseDest(ChooseDestState),
-    Customize(CustomizeState),
-    Review(CustomizeState),
-    Flashing(FlashingState),
-    FlashingCancel(FlashingFinishState),
-    FlashingFail(FlashingFailState),
-    FlashingSuccess(FlashingFinishState),
-    AppInfo(OverlayState),
+    ChooseBoard(state::ChooseBoardState),
+    ChooseOs(state::ChooseOsState),
+    ChooseDest(state::ChooseDestState),
+    Customize(state::CustomizeState),
+    Review(state::CustomizeState),
+    Flashing(state::FlashingState),
+    FlashingCancel(state::FlashingFinishState),
+    FlashingFail(state::FlashingFailState),
+    FlashingSuccess(state::FlashingFinishState),
+    AppInfo(state::OverlayState),
 }
 
 impl BBImager {
     const fn choose_board(common: BBImagerCommon) -> Self {
-        Self::ChooseBoard(ChooseBoardState {
+        Self::ChooseBoard(state::ChooseBoardState {
             common,
             selected_board: None,
         })
@@ -1068,7 +346,7 @@ impl BBImager {
 
         let (t, h) = Task::stream(s).abortable();
 
-        *self = Self::Flashing(FlashingState {
+        *self = Self::Flashing(state::FlashingState {
             is_download,
             common: state.common,
             selected_board: state.selected_board,
@@ -1082,33 +360,13 @@ impl BBImager {
 
     fn back(&mut self) {
         *self = match std::mem::take(self) {
-            Self::ChooseOs(inner) => Self::ChooseBoard(ChooseBoardState {
-                common: inner.common,
-                selected_board: Some(inner.selected_board),
-            }),
-            Self::ChooseDest(inner) => Self::ChooseOs(ChooseOsState {
-                common: inner.common,
-                selected_board: inner.selected_board,
-                pos: Vec::new(),
-                selected_image: Some(inner.selected_image),
-            }),
+            Self::ChooseOs(inner) => Self::ChooseBoard(inner.into()),
+            Self::ChooseDest(inner) => Self::ChooseOs(inner.into()),
             Self::Customize(inner) => {
                 if helpers::static_destination(inner.selected_image.1.flasher()).is_none() {
-                    Self::ChooseDest(ChooseDestState {
-                        common: inner.common,
-                        selected_board: inner.selected_board,
-                        selected_image: inner.selected_image,
-                        selected_dest: Some(inner.selected_dest),
-                        destinations: Vec::new(),
-                        filter_destination: true,
-                    })
+                    Self::ChooseDest(inner.into())
                 } else {
-                    Self::ChooseOs(ChooseOsState {
-                        common: inner.common,
-                        selected_board: inner.selected_board,
-                        pos: Vec::new(),
-                        selected_image: Some(inner.selected_image),
-                    })
+                    Self::ChooseOs(inner.into())
                 }
             }
             Self::Review(inner) => {
@@ -1121,21 +379,9 @@ impl BBImager {
                 {
                     Self::Customize(inner)
                 } else if helpers::static_destination(inner.selected_image.1.flasher()).is_none() {
-                    Self::ChooseDest(ChooseDestState {
-                        common: inner.common,
-                        selected_board: inner.selected_board,
-                        selected_image: inner.selected_image,
-                        selected_dest: Some(inner.selected_dest),
-                        destinations: Vec::new(),
-                        filter_destination: true,
-                    })
+                    Self::ChooseDest(inner.into())
                 } else {
-                    Self::ChooseOs(ChooseOsState {
-                        common: inner.common,
-                        selected_board: inner.selected_board,
-                        pos: Vec::new(),
-                        selected_image: Some(inner.selected_image),
-                    })
+                    Self::ChooseOs(inner.into())
                 }
             }
             Self::AppInfo(inner) => inner.page.into(),
@@ -1154,7 +400,7 @@ impl BBImager {
                 let selected_board = inner
                     .selected_board
                     .expect("Board should alread have been selected");
-                Self::ChooseOs(ChooseOsState {
+                Self::ChooseOs(state::ChooseOsState {
                     common: inner.common,
                     selected_board,
                     pos: Vec::with_capacity(5),
@@ -1172,7 +418,7 @@ impl BBImager {
                         &selected_image.1,
                         &dest,
                     ) {
-                        Self::Customize(CustomizeState {
+                        Self::Customize(state::CustomizeState {
                             common: inner.common,
                             selected_board: inner.selected_board,
                             selected_image,
@@ -1186,7 +432,7 @@ impl BBImager {
                             &inner.common.app_config,
                         );
 
-                        Self::Customize(CustomizeState {
+                        Self::Customize(state::CustomizeState {
                             common: inner.common,
                             selected_board: inner.selected_board,
                             selected_image,
@@ -1195,7 +441,7 @@ impl BBImager {
                         })
                     }
                 } else {
-                    Self::ChooseDest(ChooseDestState {
+                    Self::ChooseDest(state::ChooseDestState {
                         common: inner.common,
                         selected_board: inner.selected_board,
                         selected_image,
@@ -1215,7 +461,7 @@ impl BBImager {
                     &inner.selected_image.1,
                     &selected_dest,
                 ) {
-                    Self::Review(CustomizeState {
+                    Self::Review(state::CustomizeState {
                         common: inner.common,
                         selected_board: inner.selected_board,
                         selected_image: inner.selected_image,
@@ -1229,7 +475,7 @@ impl BBImager {
                         &inner.common.app_config,
                     );
 
-                    Self::Customize(CustomizeState {
+                    Self::Customize(state::CustomizeState {
                         common: inner.common,
                         selected_board: inner.selected_board,
                         selected_image: inner.selected_image,
@@ -1259,7 +505,7 @@ impl BBImager {
 
                 Task::batch([subitems_task, icons_task])
             }
-            Self::Review(inner) => match inner.customization() {
+            Self::Review(inner) => match &inner.customization {
                 helpers::FlashingCustomization::LinuxSdSysconfig(c) => {
                     let mut temp = inner
                         .app_config()
