+++ bb-imager-gui/src/ui/app_info.rs
@@ -1,9 +1,7 @@
 use iced::{Element, widget};
 
 use crate::{
-    OverlayState,
-    message::BBImagerMessage,
-    ui::helpers::{element_with_label, page_type3, selectable_text},
+    message::BBImagerMessage, state::OverlayState, ui::helpers::{element_with_label, page_type3, selectable_text}
 };
 
 const INP_BOX_WIDTH: u32 = 420;
@@ -19,14 +17,14 @@ pub(crate) fn view<'a>(state: &'a OverlayState) -> Element<'a, BBImagerMessage> 
 
 fn review_view<'a>(state: &'a OverlayState) -> Element<'a, BBImagerMessage> {
     let col = widget::column![
-        widget::image(state.common().window_icon().clone()),
+        widget::image(state.common().window_icon_handle.clone()),
         crate::constants::APP_NAME,
         crate::constants::APP_RELEASE,
         crate::constants::APP_DESC,
         widget::rule::horizontal(2),
         element_with_label(
             "Cache Directory",
-            widget::text_input(state.cache_dir_path(), state.cache_dir_path())
+            widget::text_input(&state.cache_dir, &state.cache_dir)
                 .width(INP_BOX_WIDTH)
                 .on_input(|_| BBImagerMessage::Null)
                 .into()
@@ -34,7 +32,7 @@ fn review_view<'a>(state: &'a OverlayState) -> Element<'a, BBImagerMessage> {
         widget::rule::horizontal(2),
         element_with_label(
             "Log File",
-            widget::text_input(state.log_path(), state.log_path())
+            widget::text_input(&state.log_path, &state.log_path)
                 .width(INP_BOX_WIDTH)
                 .on_input(|_| BBImagerMessage::Null)
                 .into()
